cms_admin/requirements.txt
Django>=4.2
dj-database-url>=1.0
gunicorn
psycopg2-binary
whitenoise
Pillow
django-cors-headers
django-ckeditor